# pkg/get_t/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Степан Фокин
  description: |
    API для маркетплейса с полной бизнес-логикой:
    - Управление товарами (CRUD)
    - Управление заказами с валидацией
    - Промокоды и скидки
    - Аутентификация и авторизация (JWT)
    - Ролевая модель доступа (USER, SELLER, ADMIN)
  title: Маркетплейс API
  version: 1.0.0
servers:
- description: Локальное окружение разработки
  url: http://localhost:8080/api/v1
- description: Продакшн
  url: https://api.marketplace.example.com/v1
...
paths:
  /products:
    get:
      description: |
        Возвращает список товаров с поддержкой пагинации и фильтрации.
        Доступно для всех ролей (USER, SELLER, ADMIN).
      operationId: getProducts
      parameters:
      - description: Номер страницы (начиная с 0)
        example: 0
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - description: Размер страницы
        example: 20
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - description: Фильтр по статусу товара
        example: ACTIVE
        explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/ProductStatus"
        style: form
      - description: Фильтр по категории (точное совпадение)
        example: Электроника
        explode: true
        in: query
        name: category
        required: false
        schema:
          maxLength: 100
          minLength: 1
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPageResponse"
          description: Список товаров с метаданными пагинации
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации параметров
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
      security:
      - bearerAuth: []
      summary: Получить список товаров с пагинацией и фильтрацией
      tags:
      - products
    post:
      description: |
        Создаёт новый товар в каталоге.
        - USER: запрещено (403)
        - SELLER: только свои товары (автоматически привязывается к текущему продавцу)
        - ADMIN: может создавать товары для любого продавца
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
          description: Товар успешно создан
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
      security:
      - bearerAuth: []
      summary: Создать новый товар
      tags:
      - products
  /products/{id}:
    delete:
      description: |
        Переводит товар в статус ARCHIVED (мягкое удаление).
        - USER: запрещено (403)
        - SELLER: только свои товары
        - ADMIN: любые товары
      operationId: softDeleteProduct
      parameters:
      - description: Уникальный идентификатор товара
        example: 550e8400-e29b-41d4-a716-446655440000
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Товар успешно переведён в статус ARCHIVED
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Товар не найден
      security:
      - bearerAuth: []
      summary: Мягкое удаление товара
      tags:
      - products
    get:
      description: |
        Возвращает детальную информацию о товаре.
        Доступно для всех ролей (USER, SELLER, ADMIN).
      operationId: getProductById
      parameters:
      - description: Уникальный идентификатор товара
        example: 550e8400-e29b-41d4-a716-446655440000
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
          description: Товар найден
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Товар не найден
      security:
      - bearerAuth: []
      summary: Получить товар по ID
      tags:
      - products
    put:
      description: |
        Полностью заменяет данные товара.
        - USER: запрещено (403)
        - SELLER: только свои товары
        - ADMIN: любые товары
      operationId: updateProduct
      parameters:
      - description: Уникальный идентификатор товара
        example: 550e8400-e29b-41d4-a716-446655440000
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
          description: Товар успешно обновлён
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Товар не найден
      security:
      - bearerAuth: []
      summary: Обновить товар
      tags:
      - products
  /orders:
    post:
      description: |
        Создаёт новый заказ с валидацией бизнес-правил:
        1. Ограничение частоты создания (последняя операция CREATE_ORDER)
        2. Проверка активных заказов (состояния CREATED, PAYMENT_PENDING)
        3. Проверка каталога (товары существуют и ACTIVE)
        4. Проверка остатков (stock >= quantity)
        5. Резервирование остатков
        6. Снапшот цен (price_at_order)
        7. Расчёт стоимости с учётом промокода
        8. Фиксация операции в user_operations

        - USER: разрешено
        - SELLER: запрещено (403)
        - ADMIN: разрешено
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
          description: Заказ успешно создан
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Товар не найден
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Бизнес-ошибка (активный заказ, недостаточно товара, недопу\
            стимый статус)"
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка промокода
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Превышен лимит частоты создания заказа
      security:
      - bearerAuth: []
      summary: Создать новый заказ
      tags:
      - orders
...
  /orders/{id}/cancel:
    post:
      description: |
        Отменяет заказ (переводит в состояние CANCELED).
        - USER: только свои заказы
        - SELLER: запрещено (403)
        - ADMIN: любые заказы

        Проверки:
        1. Проверка владельца
        2. Проверка состояния (только CREATED, PAYMENT_PENDING)
        3. Возврат остатков
        4. Возврат использования промокода (если был применён)
      operationId: cancelOrder
      parameters:
      - description: Уникальный идентификатор заказа
        example: 550e8400-e29b-41d4-a716-446655440001
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Заказ успешно отменён
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Заказ не найден
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недопустимый переход состояния
      security:
      - bearerAuth: []
      summary: Отменить заказ
      tags:
      - orders
  /promo-codes:
    post:
      description: |
        Создаёт новый промокод для скидок.
        - USER: запрещено (403)
        - SELLER: разрешено (свои промокоды)
        - ADMIN: разрешено
      operationId: createPromoCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoCodeCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCodeResponse"
          description: Промокод успешно создан
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
      security:
      - bearerAuth: []
      summary: Создать промокод
      tags:
      - promo-codes
...
  /auth/refresh:
    post:
      description: Обновляет access token по refresh token.
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: Токены успешно обновлены
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Невалидный или просроченный refresh token
      summary: Обновить access token
      tags:
      - auth
components:
  schemas:
    ProductStatus:
      description: Статус товара в каталоге
      enum:
      - ACTIVE
      - INACTIVE
      - ARCHIVED
      type: string
...
    RefreshTokenRequest:
      example:
        refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      properties:
        refresh_token:
          description: Refresh token для обновления access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          type: string
      required:
      - refresh_token
...
    ErrorResponse:
      example:
        error_code: PRODUCT_NOT_FOUND
        details:
          field: price
          value: "-100"
          constraint: must be > 0
        message: Product not found
        timestamp: 2024-03-04T14:30:00Z
      properties:
        error_code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          description: Человекочитаемое описание ошибки
          example: Product not found
          type: string
        details:
          additionalProperties:
            type: string
          description: Дополнительные данные об ошибке
          example:
            field: price
            value: "-100"
            constraint: must be > 0
        timestamp:
          description: Время возникновения ошибки
          example: 2024-03-04T14:30:00Z
          format: date-time
          type: string
      required:
      - error_code
      - message
      - timestamp
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: |-
        JWT access token для аутентификации.
        Формат: "Bearer {access_token}"

        Роли в токене (claim "role"):
        - USER: обычный покупатель
        - SELLER: продавец (может управлять товарами)
        - ADMIN: администратор (полный доступ)
      scheme: bearer
      type: http